+++ ruoyi-system/src/main/java/com/ruoyi/system/domain/CqieSportCalendar.java
@@ -14,7 +14,7 @@ public class CqieSportCalendar {
 
     /** 运动距离 */
     @Excel(name = "运动距离")
-    private Long distance;
+    private Double distance;
 
     /** 运动轨迹 */
     @Excel(name = "运动轨迹")
@@ -22,15 +22,15 @@ public class CqieSportCalendar {
 
     /** 运动时长 */
     @Excel(name = "运动时长")
-    private Long duration;
+    private Double duration;
 
     /** 消耗卡路里 */
     @Excel(name = "消耗卡路里")
-    private Long calorie;
+    private Double calorie;
 
     /** 平均配速（分钟/公里） */
     @Excel(name = "平均配速", readConverterExp = "分钟/公里")
-    private Long distribution;
+    private Double distribution;
 
     /** 开跑时间 */
     @Excel(name = "开跑时间", width = 30, dateFormat = "yyyy-MM-dd HH:mm:ss")
@@ -43,11 +43,11 @@ public class CqieSportCalendar {
     /** 数据状态 */
     private  Long status;
 
-    public Long getDistance() {
+    public Double getDistance() {
         return distance;
     }
 
-    public void setDistance(Long distance) {
+    public void setDistance(Double distance) {
         this.distance = distance;
     }
 
@@ -59,27 +59,27 @@ public class CqieSportCalendar {
         this.pathLine = pathLine;
     }
 
-    public Long getDuration() {
+    public Double getDuration() {
         return duration;
     }
 
-    public void setDuration(Long duration) {
+    public void setDuration(Double duration) {
         this.duration = duration;
     }
 
-    public Long getCalorie() {
+    public Double getCalorie() {
         return calorie;
     }
 
-    public void setCalorie(Long calorie) {
+    public void setCalorie(Double calorie) {
         this.calorie = calorie;
     }
 
-    public Long getDistribution() {
+    public Double getDistribution() {
         return distribution;
     }
 
-    public void setDistribution(Long distribution) {
+    public void setDistribution(Double distribution) {
         this.distribution = distribution;
     }
